move createRowEntries into descriptions, verbForm into verbs
