+++ javascript/react/src/components/StepDescription/createRowEntries.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type { DescriptionRow, StepDescriptionProps } from '../types.js'
+import type { DescriptionRow, StepDescriptionProps } from '../../types.js'
 
 // this is the maximum rows a verb description can show.
 // it is based on the max core requirement, currently held by BIN
