+++ javascript/react/src/components/StepDescription/verbDescriptions/UnhotDescription.tsx
@@ -7,7 +7,7 @@ import { memo, useMemo } from 'react'
 
 import { EMPTY_ARRAY } from '../../../empty.js'
 import type { StepDescriptionProps } from '../../../types.js'
-import { createRowEntries } from '../../createRowEntries.js'
+import { createRowEntries } from '../createRowEntries.js'
 import { VerbDescription } from './VerbDescription.js'
 
 export const UnhotDescription: React.FC<StepDescriptionProps<UnhotArgs>> = memo(
