+++ javascript/react/src/components/verbs/Bin/Bin.base.tsx
@@ -14,7 +14,7 @@ import {
 	enumDropdown,
 	FormInputType,
 	VerbForm,
-} from '../../verbForm/index.js'
+} from '../verbForm/index.js'
 
 /**
  * Provides inputs for a binning step.
