+++ javascript/react/src/components/verbs/Join/Join.base.tsx
@@ -15,7 +15,7 @@ import {
 	joinInputs,
 	tableDropdown,
 	VerbForm,
-} from '../../verbForm/index.js'
+} from '../verbForm/index.js'
 
 /**
  * Provides inputs for a Join step.
