+++ javascript/react/src/components/verbs/Lookup/Lookup.base.tsx
@@ -14,7 +14,7 @@ import {
 	joinInputs,
 	tableDropdown,
 	VerbForm,
-} from '../../verbForm/index.js'
+} from '../verbForm/index.js'
 /**
  * Provides inputs for a Lookup step.
  */
