+++ javascript/react/src/components/verbs/OneHot/OneHot.base.tsx
@@ -8,11 +8,7 @@ import cloneDeep from 'lodash-es/cloneDeep.js'
 import { memo, useCallback, useMemo } from 'react'
 
 import type { StepComponentBaseProps } from '../../../types.js'
-import {
-	type FormInput,
-	FormInputType,
-	VerbForm,
-} from '../../verbForm/index.js'
+import { type FormInput, FormInputType, VerbForm } from '../verbForm/index.js'
 
 /**
  * Provides inputs for a OneHot step.
