+++ javascript/react/src/components/verbs/verbForm/NumberSpinnerInput.types.ts
