+++ javascript/react/src/components/verbs/verbForm/VerbForm.tsx
@@ -7,7 +7,7 @@ import { Expando } from '@essex/components'
 import { memo, useMemo } from 'react'
 import { Case, Switch, When } from 'react-if'
 
-import type { StepChangeFunction } from '../../types.js'
+import type { StepChangeFunction } from '../../../types.js'
 import { CheckboxInput } from './CheckboxInput.js'
 import type { CheckboxFormInput } from './CheckboxInput.types.js'
 import { ComboBoxInput } from './ComboBoxInput.js'
@@ -20,7 +20,7 @@ import { SingleChoiceInput } from './SingleChoiceInput.js'
 import type { SingleChoiceFormInput } from './SingleChoiceInput.types.js'
 import { TextInput } from './TextInput.js'
 import type { TextFormInput } from './TextInput.types.js'
-import type { FormInput} from './types.js';
+import type { FormInput } from './types.js'
 import { FormInputType } from './types.js'
 import { useSortedInputs } from './VerbForm.hooks.js'
 import { Container, InputsBlock, Row } from './VerbForm.styles.js'
